refactor(components): move components to `src/components`
Move global and shareable components from `src/app/[route]/_components/*` to `src/components`.
