+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
+import Footer from '@components/Footer';
+import Header from '@components/Header';
 import type { Metadata } from 'next';
 import type { JSX } from 'react';
 import { geistMono, geistSans } from '../consts/geist-fonts.const';
-import Footer from './_components/Footer';
-import Header from './_components/Header';
 import './globals.css';
 
 export const metadata: Metadata = {
